Nominal quines in reducer style as suggested by @webyrd.
In small step, values are wrapped in (quote ...) so they do not get
reduced. Hence, a true quine is a program which eventually reduces to
its quote.

In both small step and big step, the reducers are able to accept
quines, but they cannot generate them (yet!). Not sure if it's a bug
in core.logic.nominal, or if the programs need to be fixed.

core.logic.nominal is available at:
https://github.com/namin/core.logic/compare/master...nominal-base